+++ Development/Source/tools.hpp
@@ -14,7 +14,7 @@ public:
 	static const string getRootDirectory();
 	static const int getRandomInt(int min, int max);
 	static const float getRandomFloat(float min, float max);
-	static const unsigned getTimeSinceEpochMS();
+	static const long long getTimeSinceEpochMS();
 
 private:
 	static inline std::mt19937 _generator = std::mt19937(static_cast<unsigned int>(time(nullptr)));
